refactor: 移除更新模块对用户数据的依赖
